+++ server/src/main/java/org/allaymc/server/entity/component/EntityFallingBlockBaseComponentImpl.java
@@ -114,7 +114,7 @@ public class EntityFallingBlockBaseComponentImpl extends EntityBaseComponentImpl
     }
 
     @Override
-    public float getBaseOffset() {
+    public float getNetworkOffset() {
         return 0.49f;
     }
 }